pyproject.toml
[project]
name = "crying-baby-audio-scrub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
